# openapi.yaml
openapi: 3.0.3
info:
  title: FLYFOX AI - NQBA Ecosystem API
  version: "2.0.0"
  description: |
    Complete API for FLYFOX AI's Neuromorphic Quantum Business Architecture (NQBA) ecosystem.
    
    ## Core Components
    
    ### 🎓 FLYFOX Academy
    - Training program enrollment and progress tracking
    - Certification issuance and verification
    - Developer program access and marketplace integration
    
    ### 🔄 Cross-Agent Orchestrator
    - Workflow execution and management
    - Agent registration and capability discovery
    - Cross-agent communication and collaboration
    
    ### ⚡ Quantum Integration Hub (QIH)
    - Quantum job submission and management
    - Optimization problem solving
    - Performance benchmarking and analytics
    
    ### 🏢 Business Units
    - FLYFOX AI (Energy Optimization)
    - Goliath Trade (Financial Intelligence)
    - Sigma Select (Lead Scoring & Sales)
    
    ### 🔐 Authentication & Entitlements
    - JWT-based authentication
    - Role-based access control (RBAC)
    - Feature gating and tier management
    
    ## Technology Stack
    - **Backend**: FastAPI + Python 3.11+
    - **Quantum**: Dynex SDK integration
    - **Database**: PostgreSQL + Redis
    - **Monitoring**: OpenTelemetry + Prometheus + Grafana
    - **Deployment**: Docker + Kubernetes + AWS EKS
    
    ## Getting Started
    1. Obtain API credentials from FLYFOX AI
    2. Authenticate using JWT tokens
    3. Explore available endpoints and capabilities
...
    
...

...
servers:
...
paths:
...
  /academy/learning-paths:
    get:
      summary: Get Learning Paths
      description: Retrieve all available learning paths for FLYFOX Academy
      tags: [FLYFOX Academy]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of learning paths
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LearningPathResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /academy/enroll:
    post:
      summary: Enroll in Learning Path
      description: Enroll a user in a specific learning path
      tags: [FLYFOX Academy]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRequest'
      responses:
        '201':
          description: Successfully enrolled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  path_id:
                    type: string
                  tier:
                    type: string
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /academy/progress:
    get:
      summary: Get User Progress
      description: Retrieve learning progress for the authenticated user
      tags: [FLYFOX Academy]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User progress data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProgressResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

    put:
      summary: Update Progress
      description: Update user progress in a learning path
      tags: [FLYFOX Academy]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressUpdateRequest'
      responses:
        '200':
          description: Progress updated successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /academy/certificate:
    post:
      summary: Request Certificate
      description: Request a certificate for a completed learning path
      tags: [FLYFOX Academy]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateRequest'
      responses:
        '201':
          description: Certificate issued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  certificate_id:
                    type: string
                  path_id:
                    type: string
                  level:
                    type: string
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /academy/certificates/{certificate_id}:
    get:
      summary: Get Certificate
      description: Retrieve certificate details by ID
      tags: [FLYFOX Academy]
      security:
        - bearerAuth: []
      parameters:
        - name: certificate_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Certificate details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Certificate not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
  schemas:
...
    AcademyTier:
      type: string
      enum: [free, student, professional, enterprise, partner]
      description: Academy access tiers for different user types
    CertificationLevel:
      type: string
      enum: [bronze, silver, gold, platinum, master]
      description: Certification levels for quantum AI expertise
    EnrollmentRequest:
      type: object
      required:
        - path_id
        - tier
      properties:
        path_id:
          type: string
          description: ID of the learning path to enroll in
        tier:
          $ref: '#/components/schemas/AcademyTier'
    ProgressUpdateRequest:
      type: object
      required:
        - path_id
        - module_completion
        - time_spent
      properties:
        path_id:
          type: string
          description: ID of the learning path
        module_completion:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: Module completion percentage
        time_spent:
          type: integer
          minimum: 0
          description: Time spent in minutes
    CertificateRequest:
      type: object
      required:
        - path_id
        - level
      properties:
        path_id:
          type: string
          description: ID of the completed learning path
        level:
          $ref: '#/components/schemas/CertificationLevel'
    LearningPathResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        difficulty:
          type: string
        estimated_hours:
          type: integer
        modules:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              description:
                type: string
              duration_minutes:
                type: integer
              content_type:
                type: string
    UserProgressResponse:
      type: object
      properties:
        path_id:
          type: string
        completion_percentage:
          type: number
          format: float
        time_spent_minutes:
          type: integer
        assessments_passed:
          type: integer
        total_assessments:
          type: integer
        certificate_earned:
          type: string
          nullable: true
    CertificateResponse:
      type: object
      properties:
        certificate_id:
          type: string
        path_id:
          type: string
        level:
          type: string
        issued_date:
          type: string
          format: date-time
        expiry_date:
          type: string
          format: date-time
        metadata:
          type: object
...